Use size_t for font/typeface return values (to match span's)
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1004919
Commit-Queue: Mike Reed <mike@reedtribe.org>

+++ gm/rsxtext.cpp
@@ -42,7 +42,7 @@ private:
         std::copy(glyphs, glyphs + glyph_count, buf.glyphs);
 
         float x = 0;
-        for (int i = 0; i < glyph_count; ++i) {
+        for (size_t i = 0; i < glyph_count; ++i) {
             buf.xforms()[i] = {
                 1, 0,
                 x, 0,
